backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.9.0
motor==3.6.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
apscheduler==3.10.4
openai>=1.10.0
python-multipart==0.0.6
email-validator==2.1.0
chromadb>=0.4.22
PyPDF2==3.0.1
langchain>=0.1.10
langchain-openai>=0.0.5
langchain-community>=0.0.24
duckduckgo-search>=6.2.6
beautifulsoup4>=4.12.3
